setup: Specify `gevent` as a hard dependency.
Fixes #72.

`gevent` is used in code directly, but wasn't listed as a dep.

This used to work before, since `populus` (another dep of EAC)
had a hard dep on `gevent`. It has become "extra" there - see
https://github.com/pipermerriam/populus/pull/218

Following the docs on EAC instance deployment otherwise
results in a client that's unable to start.