pypi: fix typo, do include the contracts' ABI JSON in the package.
Fixes #74.

The MANIFEST.in had a typo and didn't include the ABI description
JSON for EAC's contracts in the PyPI package.

It also seems that the version tag had a typo, with an unlikely
leap from v0.7.0 to v8.0.0b1. This can't be fixed in a PR, and
will require yanking the v8.0.0b1 release from PyPI, together
with a loud communique to everyone who happened to install that
version.